commons: add deserializer helpers
